Module artemis_normalized_cache::cache_exchange

Contains the exchange factory and implementation. The factory is the only thing needed for most users and is reexported from the root.

Structs

NormalizedCacheExchange
The normalized cache exchange. This will store normalized queries by unique ID.

NormalizedCacheImpl
The implementation of the normalized cache. Exposed in case someone needs it, but most users shouldn't.